requirements.txt
flask
flask-restx
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
werkzeug
psycopg2-binary
pytest
pytest-flask
pytest-cov
requests